Create Question1-wordEmbeddings
Question
There is a list of phrases given. (meaning of some words in a dictionary)
We need to find the word embeddings of each phrase. (Need to look at how to represent a phrase as a vector as words itself are a vector).
A word will be given as an input. Need to find euclidean distance difference (some suitable algorithm). And check the difference between the vector of the given input word from the list of the phrases (calculated in question part 1).  Afterwards, return the word/ phrase vector that is closest to the given word with the input word vector.

Problem 1: Convert a list of english phrases to vector representations. Ensure adequate preprocessing to get rid of stop words, lemmatization , tfidf etc. Next given a new word, map it to the phrase that’s closest to it in the vector space. Define appropriate functions to achieve this and write a program in python 3.11 to demonstrate this. Start with a short list of 10 - 20 phrases that you make up (or get from some website). The actual phrase list will be shared with you in a few hours.