Create Q1-WordEmbeddings-Updated2
Updates

Problem 1: Convert a list of english phrases to vector representations. Ensure adequate preprocessing to get rid of stop words, lemmatization , tfidf etc. Next given a new word, map it to the phrase that’s closest to it in the vector space. Define appropriate functions to achieve this and write a program in python 3.11 to demonstrate this. Start with a short list of 10 - 20 phrases that you make up (or get from some website). The actual phrase list will be shared with you in a few hours.

Feedback from the meeting on May 25

f1) Clean up your python notebook where you are presenting your work for review so that it is readable and does not have code that is not relevant to the problem
a1) Done where required

f2) Create the input data structure programmatically. Ensure that you do the requires text preprocessing (removal of brackets, stop words, high freq words, carry out lemmatization, convert numerals to text …) 
a2) Done

f3) The problem is to find the phrase that is the closest match to a given word, not the token that is the closest match to a given word. Please solve the problem that we defined (first para of this document). 
a3) Done

f4) Only write code when you are sure about what it is meant to do. 
a4) OK
 
f5) Do not include code just because it works but the functionality of which is not clear to you
a5) OK

f6) Deliver the solution to me in the form of a function that I can import into my own program and use.
a6) Two programs written. This one is not having function. Program with function to follow.