Create Q1_WordEmbeddingsUpdated3
Update 29-05-2023
f6 ) Deliver the solution to me in the form of a function that I can import into my own program and use.
a6) We have arrived at a legitimate program that has been updated in function form. Complete re-work has been done to solve the problem by using a different technology method detailed in the file uploaded on GitHub.